split auto_config auto_utils auto_env
